+++ frontend/components/CenteredDiv.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export default styled.div`
+  min-height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
