+++ frontend/types/handleRegister.ts
@@ -1,0 +1,1 @@
+export type HandleRegister = (success: boolean) => void;
