+++ frontend/types/index.ts
@@ -1,1 +1,2 @@
 export * from './handleLogin';
+export * from './handleRegister';
